Reordered methods, renamed hash method and added documentation to Crypto class methods
